Rename dashboard publishing service to build enqueuer
dashboard_publishing.py's DashboardPublishingService and the
try-enqueue helpers in dashboard_trigger.py both existed only to put a
dashboard_build job on the queue — the former raises on failure, the
latter swallows it. Merge them into services/dashboard/enqueue.py so
the raise-vs-swallow variants sit side by side, and rename the class
to DashboardBuildEnqueuer so the name no longer collides with the
render-pipeline DashboardPublisher.

- src/docverse/services/dashboard_publishing.py ->
  src/docverse/services/dashboard/enqueue.py
- Absorb src/docverse/services/dashboard_trigger.py's two
  try_enqueue_dashboard_build_by_{slug,id} free functions into the
  new enqueue.py, then delete the trigger module.
- Rename DashboardPublishingService -> DashboardBuildEnqueuer.
- Rename Factory.create_dashboard_publishing_service ->
  create_dashboard_build_enqueuer; update handler and worker call
  sites that imported from services.dashboard_trigger to import from
  services.dashboard.enqueue instead.
- Update the enqueue-triggers test imports and monkeypatch targets.
