+++ src/docverse/handlers/orgs/editions.py
@@ -24,7 +24,7 @@ from docverse.handlers.params import (
     OrgSlugParam,
     ProjectSlugParam,
 )
-from docverse.services.dashboard_trigger import (
+from docverse.services.dashboard.enqueue import (
     try_enqueue_dashboard_build_by_slug,
 )
 from docverse.storage.pagination import (
