+++ src/docverse/services/dashboard/enqueue.py
@@ -2,6 +2,8 @@
 
 from __future__ import annotations
 
+from typing import TYPE_CHECKING
+
 import structlog
 
 from docverse.client.models.queue_enums import JobKind
@@ -13,10 +15,19 @@ from docverse.storage.project_store import ProjectStore
 from docverse.storage.queue_backend import QueueBackend
 from docverse.storage.queue_job_store import QueueJobStore
 
-__all__ = ["DashboardPublishingService"]
+if TYPE_CHECKING:
+    from sqlalchemy.ext.asyncio import AsyncSession
+
+    from docverse.factory import Factory
+
+__all__ = [
+    "DashboardBuildEnqueuer",
+    "try_enqueue_dashboard_build_by_id",
+    "try_enqueue_dashboard_build_by_slug",
+]
 
 
-class DashboardPublishingService:
+class DashboardBuildEnqueuer:
     """Enqueue dashboard render jobs.
 
     This is the thin slice in the MVP — it creates the ``QueueJob`` row
@@ -136,3 +147,57 @@ class DashboardPublishingService:
             )
             results.append((project, queue_job))
         return results
+
+
+async def try_enqueue_dashboard_build_by_slug(
+    *,
+    factory: Factory,
+    session: AsyncSession,
+    logger: structlog.stdlib.BoundLogger,
+    org_slug: str,
+    project_slug: str,
+) -> None:
+    """Enqueue one ``dashboard_build`` job in its own transaction.
+
+    Exceptions are logged but never re-raised, so the caller's flow is
+    not broken by an enqueue failure. The enqueue runs in a freshly
+    started transaction on ``session`` — the caller must have already
+    committed any work it wants persisted.
+    """
+    try:
+        async with session.begin():
+            service = factory.create_dashboard_build_enqueuer()
+            await service.enqueue_for_project_slug(
+                org_slug=org_slug, project_slug=project_slug
+            )
+            await session.commit()
+    except Exception:
+        logger.exception(
+            "Failed to enqueue dashboard_build",
+            org_slug=org_slug,
+            project_slug=project_slug,
+        )
+
+
+async def try_enqueue_dashboard_build_by_id(
+    *,
+    factory: Factory,
+    session: AsyncSession,
+    logger: structlog.stdlib.BoundLogger,
+    org_id: int,
+    project_id: int,
+) -> None:
+    """ID-based variant of :func:`try_enqueue_dashboard_build_by_slug`."""
+    try:
+        async with session.begin():
+            service = factory.create_dashboard_build_enqueuer()
+            await service.enqueue_for_project(
+                org_id=org_id, project_id=project_id
+            )
+            await session.commit()
+    except Exception:
+        logger.exception(
+            "Failed to enqueue dashboard_build",
+            org_id=org_id,
+            project_id=project_id,
+        )
